+++ app/package.json
@@ -1,7 +1,7 @@
 {
   "name": "root",
   "version": "1.0.0",
-  "buildVersion": "cdefbae",
+  "buildVersion": "320d92c",
   "private": true,
   "devDependencies": {
     "lerna": "^3.20.2",
