+++ lib/core/src/display/scene.rs
@@ -32,6 +32,7 @@ use crate::system::web::resize_observer::ResizeObserver;
 use crate::system::web::StyleSetter;
 use crate::system::web;
 use crate::system::web::NodeInserter;
+use crate::system::web::dom;
 
 use wasm_bindgen::prelude::Closure;
 use wasm_bindgen::JsValue;
@@ -51,142 +52,6 @@ pub enum Error {
 
 
 
-// =============
-// === Shape ===
-// =============
-
-// === Shape ===
-
-use shapely::shared;
-
-shared! { Shape
-/// Contains information about DOM element shape and provides utils for querying the size both in
-/// DOM pixel units as well as device pixel units.
-#[derive(Debug)]
-##[derive(Clone,Copy)]
-pub struct ShapeData {
-    width       : f32,
-    height      : f32,
-    pixel_ratio : f32
-}
-
-impl {
-    /// Constructor.
-    pub fn new() -> Self {
-        let width       = 0.0;
-        let height      = 0.0;
-        let pixel_ratio = web::device_pixel_ratio().unwrap_or(1.0) as f32;
-        Self {width,height,pixel_ratio}
-    }
-
-    /// Getter.
-    pub fn width(&self) -> f32 {
-        self.width
-    }
-
-    /// Getter.
-    pub fn height(&self) -> f32 {
-        self.height
-    }
-
-    /// Getter.
-    pub fn pixel_ratio(&self) -> f32 {
-        self.pixel_ratio
-    }
-
-    /// Dimension setter in DOM pixel units. Use `device_pixels` to switch to device units.
-    pub fn set(&mut self, width:f32, height:f32) {
-        self.width  = width;
-        self.height = height;
-    }
-
-    /// Sets the size to the size of the provided element. This operation is slow as it causes
-    /// reflow.
-    pub fn set_from_element_with_reflow(&mut self, element:&HtmlElement) {
-        let bbox   = element.get_bounding_client_rect();
-        let width  = bbox.width()  as f32;
-        let height = bbox.height() as f32;
-        self.set(width,height);
-    }
-}}
-
-impl ShapeData {
-    /// Switched to device pixel units. On low-dpi screens device pixels map 1:1 with DOM pixels.
-    /// On high-dpi screens, a single device pixel is often mapped to 2 or 3 DOM pixels.
-    pub fn device_pixels(&self) -> Self {
-        let width  = self.width  * self.pixel_ratio;
-        let height = self.height * self.pixel_ratio;
-        Self {width,height,..*self}
-    }
-}
-
-impl Shape {
-    pub fn from_element_with_reflow(element:&HtmlElement) -> Self {
-        let mut this = Self::default();
-        this.set_from_element_with_reflow(element);
-        this
-    }
-
-    pub fn current(&self) -> ShapeData {
-        *self.rc.borrow()
-    }
-}
-
-impl Default for Shape {
-    fn default() -> Self {
-        Self::new()
-    }
-}
-
-impl Default for ShapeData {
-    fn default() -> Self {
-        Self::new()
-    }
-}
-
-
-
-// ====================
-// === SizedElement ===
-// ====================
-
-/// A wrapper for `HtmlElement` or anything which derefs to it. It tracks the element size without
-/// causing browser reflow.
-#[derive(Debug,Shrinkwrap)]
-pub struct SizedElement<T=web_sys::HtmlElement> {
-    #[shrinkwrap(main_field)]
-    dom       : T,
-    shape     : Shape,
-    observer  : ResizeObserver,
-    on_resize : Rc<RefCell<CallbackRegistry1<ShapeData>>>,
-}
-
-impl<T> SizedElement<T> {
-    /// Constructor.
-    pub fn new(dom:&T) -> Self
-    where T : Clone + AsRef<JsValue> + Into<web_sys::HtmlElement> {
-        let dom          = dom.clone();
-        let html_element = dom.clone().into();
-        let shape        = Shape::from_element_with_reflow(&html_element);
-        let on_resize    = Rc::new(RefCell::new(CallbackRegistry1::default()));
-        let callback     = Closure::new(enclose!((shape,on_resize) move |width, height| {
-            shape.set(width as f32,height as f32);
-            on_resize.borrow_mut().run_all(&shape.current())
-        }));
-        let observer = ResizeObserver::new(dom.as_ref(),callback);
-        Self {dom,shape,observer,on_resize}
-    }
-
-    /// Attach a new callback which will fire whenever the object will be resized.
-    pub fn on_resize<F:CallbackMut1Fn<ShapeData>>(&self, callback:F) -> CallbackHandle {
-        self.on_resize.borrow_mut().add(callback)
-    }
-
-    /// Get the current shape of the object.
-    pub fn shape(&self) -> &Shape {
-        &self.shape
-    }
-}
 
 
 
@@ -217,7 +82,7 @@ struct Mouse {
 }
 
 impl Mouse {
-    pub fn new(shape:&Shape, variables:&UniformScope) -> Self {
+    pub fn new(shape:&dom::Shape, variables:&UniformScope) -> Self {
 
         let empty_hover_ids = Vector4::<u32>::new(0,0,0,0);
         let position        = variables.add_or_panic("mouse_position",Vector2::new(0,0));
@@ -236,7 +101,7 @@ impl Mouse {
         let on_move_handle  = mouse_manager.on_move.add(move |event:&mouse::event::OnMove| {
             let pixel_ratio = shape_ref.pixel_ratio() as i32;
             let screen_x    = event.offset_x();
-            let screen_y    = shape_ref.current().height as i32 - event.offset_y();
+            let screen_y    = shape_ref.current().height() as i32 - event.offset_y();
             let canvas_x    = pixel_ratio * screen_x;
             let canvas_y    = pixel_ratio * screen_y;
             position_ref.set(Vector2::new(canvas_x,canvas_y))
@@ -289,7 +154,7 @@ impl Mouse {
 #[derive(Debug)]
 pub struct Dom {
     /// Root DOM element of the scene.
-    pub root : SizedElement<web_sys::HtmlDivElement>,
+    pub root : dom::WithKnownShape<web_sys::HtmlDivElement>,
     /// Layers of the scene.
     pub layers : Layers,
 }
@@ -302,11 +167,11 @@ impl Dom {
         root.set_style_or_panic("height"  , "100vh");
         root.set_style_or_panic("width"   , "100vw");
         root.set_style_or_panic("display" , "block");
-        let root = SizedElement::new(&root);
+        let root = dom::WithKnownShape::new(&root);
         Self {root,layers}
     }
 
-    pub fn shape(&self) -> &Shape {
+    pub fn shape(&self) -> &dom::Shape {
         self.root.shape()
     }
 }
@@ -363,10 +228,8 @@ pub struct SceneData {
     symbols        : SymbolRegistry,
     symbols_dirty  : SymbolRegistryDirty,
     camera         : Camera2d,
-//    shape          : Shape,
     shape_dirty    : ShapeDirty,
     logger         : Logger,
-//    listeners      : Listeners,
     variables      : UniformScope,
     pipeline       : RenderPipeline,
     composer       : RenderComposer,
@@ -376,8 +239,6 @@ pub struct SceneData {
     zoom_callback  : CallbackHandle,
     mouse          : Mouse,
     on_resize      : CallbackHandle,
-//    #[derivative(Debug="ignore")]
-//    on_resize2: Option<Box<dyn Fn(&Shape)>>,
 }
 
 impl {
@@ -400,21 +261,14 @@ impl {
         let variables       = UniformScope::new(logger.sub("global_variables"),&context);
         let symbols         = SymbolRegistry::new(&variables,&stats,&context,sub_logger,on_change);
 
-        println!("--- 1");
         dom.shape().set_from_element_with_reflow(&dom.root);
-        println!("--- 2");
 
-        println!("--- 3 {:?}",dom.shape() );
-
-//        let shape           = Shape::from_element_with_reflow(&dom.root);
         let screen_shape    = dom.shape().current();
-        let width           = screen_shape.width;
-        let height          = screen_shape.height;
-//        let listeners       = Self::init_listeners(&logger,&dom.layers.canvas,&shape,&shape_dirty);
+        let width           = screen_shape.width();
+        let height          = screen_shape.height();
         let symbols_dirty   = dirty_flag;
         let camera          = Camera2d::new(logger.sub("camera"),width,height);
         let zoom_uniform    = variables.add_or_panic("zoom", 1.0);
-//        let on_resize2       = default();
         let stats           = stats.clone();
         let pixel_ratio     = variables.add_or_panic("pixel_ratio", dom.shape().pixel_ratio());
         let mouse           = Mouse::new(&dom.shape(),&variables);
@@ -423,12 +277,9 @@ impl {
             move |zoom:&f32| zoom_uniform_cp.set(*zoom)
         );
 
-        println!("x1");
-        let on_resize = dom.root.on_resize(enclose!((shape_dirty) move |_:&ShapeData| {
-            println!("resized callback");
+        let on_resize = dom.root.on_resize(enclose!((shape_dirty) move |_:&dom::ShapeData| {
             shape_dirty.set();
         }));
-        println!("x2");
 
         context.enable(Context::BLEND);
         // To learn more about the blending equations used here, please see the following articles:
@@ -439,8 +290,8 @@ impl {
                                         , Context::ONE , Context::ONE_MINUS_SRC_ALPHA );
 
         let pipeline = default();
-        let width    = dom.shape().current().device_pixels().width  as i32;
-        let height   = dom.shape().current().device_pixels().height as i32;
+        let width    = dom.shape().current().device_pixels().width()  as i32;
+        let height   = dom.shape().current().device_pixels().height() as i32;
         let composer = RenderComposer::new(&pipeline,&context,&variables,width,height);
 
         Self { pipeline,composer,display_object,dom,context,symbols,camera,symbols_dirty,shape_dirty
@@ -482,8 +333,8 @@ impl {
     }
 
     pub fn init_composer(&mut self) {
-        let width    = self.dom.shape().current().device_pixels().width  as i32;
-        let height   = self.dom.shape().current().device_pixels().height as i32;
+        let width    = self.dom.shape().current().device_pixels().width()  as i32;
+        let height   = self.dom.shape().current().device_pixels().height() as i32;
         self.composer = RenderComposer::new(&self.pipeline,&self.context,&self.variables,width,height);
     }
 
@@ -510,7 +361,7 @@ impl {
             if self.shape_dirty.check_all() {
                 let screen = self.dom.shape().current();
                 self.resize_canvas(&self.dom.shape());
-                self.camera.set_screen(screen.width, screen.height);
+                self.camera.set_screen(screen.width(), screen.height());
                 self.init_composer();
                 self.shape_dirty.unset_all();
             }
@@ -600,36 +451,16 @@ impl Scene {
 }
 
 impl SceneData {
-    /// Initialize all listeners and attach them to DOM elements.
-    fn init_listeners
-    (logger:&Logger, canvas:&web_sys::HtmlCanvasElement, shape:&Shape, dirty:&ShapeDirty)
-    -> Listeners {
-        let logger = logger.clone();
-        let shape  = shape.clone();
-        let dirty  = dirty.clone();
-        let on_resize = Closure::new(move |width, height| {
-            group!(logger, "Resize observer event.", {
-                shape.set(width as f32,height as f32);
-                dirty.set();
-            })
-        });
-        let resize = ResizeObserver::new(canvas,on_resize);
-        Listeners {resize}
-    }
-
-
-
     /// Resize the underlying canvas. This function should rather not be called
     /// directly. If you want to change the canvas size, modify the `shape` and
     /// set the dirty flag.
-    fn resize_canvas(&self, shape:&Shape) {
+    fn resize_canvas(&self, shape:&dom::Shape) {
         let screen = shape.current();
         let canvas = shape.current().device_pixels();
-        self.logger.group(fmt!("Resized to {}px x {}px.", screen.width, screen.height), || {
-            self.dom.layers.canvas.set_attribute("width",  &canvas.width.to_string()).unwrap();
-            self.dom.layers.canvas.set_attribute("height", &canvas.height.to_string()).unwrap();
-            self.context.viewport(0,0,canvas.width as i32, canvas.height as i32);
-//            self.on_resize2.iter().for_each(|f| f(shape));
+        self.logger.group(fmt!("Resized to {}px x {}px.", screen.width(), screen.height()), || {
+            self.dom.layers.canvas.set_attribute("width",  &canvas.width().to_string()).unwrap();
+            self.dom.layers.canvas.set_attribute("height", &canvas.height().to_string()).unwrap();
+            self.context.viewport(0,0,canvas.width() as i32, canvas.height() as i32);
         });
     }
 }
