+++ lib/core/src/system/web.rs
@@ -4,3 +4,6 @@ pub mod dom;
 pub mod text_input;
 
 pub use basegl_system_web::*;
+pub use web_sys::HtmlElement;
+pub use web_sys::HtmlDivElement;
+pub use wasm_bindgen::JsValue;
