+++ lib/core/src/system/web/dom.rs
@@ -1,5 +1,7 @@
 //! This module exposes CSS3D rendering system and other DOM helpers.
 
 pub mod html;
+pub mod shape;
 
+pub use shape::*;
 pub use basegl_system_web::dom::*;
